Refactor Tests: yield_resources to its own file and trade_possibilities to actions
